refactor(entity): change table column width type from double to int
Convert widthFlex values from double to int type for consistency and simplify table layout configuration. The conversion is handled with toDouble() where needed.

+++ lib/src/app/view/widget/entity_data_table.dart
@@ -174,9 +174,9 @@ class _MenuDataTableCustomState extends State<MenuDataTableCustom> {
                 );
               }
 
-              final width = fieldsValue[index] as double?;
+              final width = fieldsValue[index] as int?;
               return DTColumn(
-                widthFlex: width ?? 1,
+                widthFlex: (width ?? 1).toDouble(),
                 head: DTHead(
                   backendKeySort: field.reference,
                   label: field.label,
